playground/requirements.txt
pygame
PyOpenGL